New initial partition implementation
Added initial partition implementation for randomly selecting 52 starting cbgs and growing them out using bfs
