+++ final_shapefile_without_water/final_shapefile_without_water.cpg
@@ -1,0 +1,1 @@
+UTF-8
